Replace all occurences of itemId from int to long

+++ BunqSdk/Context/ApiContext.cs
@@ -31,12 +31,12 @@ namespace Bunq.Sdk.Context
         /// <summary>
         /// Measure of any time unit when none of it is needed.
         /// </summary>
-        private const int TIME_UNIT_COUNT_NONE = 0;
+        private const long TIME_UNIT_COUNT_NONE = 0;
 
         /// <summary>
         /// Minimum time to session expiry not requiring session reset.
         /// </summary>
-        private const int TIME_TO_SESSION_EXPIRY_MINIMUM_SECONDS = 30;
+        private const long TIME_TO_SESSION_EXPIRY_MINIMUM_SECONDS = 30;
 
         /// <summary>
         /// Default path to store the serialized context.
@@ -46,7 +46,7 @@ namespace Bunq.Sdk.Context
         /// <summary>
         /// Dummy ID to pass to Session endpoint.
         /// </summary>
-        private const int SESSION_ID_DUMMY = 0;
+        private const long SESSION_ID_DUMMY = 0;
 
         /// <summary>
         /// Encoding of the serialized context.
