+++ BunqSdk/Model/Core/Id.cs
@@ -4,9 +4,9 @@ namespace Bunq.Sdk.Model.Core
 {
     public class Id
     {
-        [JsonProperty(PropertyName = "id")] public int IdInt { get; private set; }
+        [JsonProperty(PropertyName = "id")] public long IdInt { get; private set; }
 
-        public Id(int idInt)
+        public Id(long idInt)
         {
             IdInt = idInt;
         }
